create new company

+++ src/new-company/new-company.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, UnauthorizedException } from "@nestjs/common";
+import { BadRequestException, Injectable, UnauthorizedException } from "@nestjs/common";
 import { CreateNewCompanyDto } from "./dto/create-new-company.dto";
 import { UpdateNewCompanyDto } from "./dto/update-new-company.dto";
 import { InjectRepository } from "@nestjs/typeorm";
@@ -11,6 +11,7 @@ import { UsersRoles } from "src/usersCompanies/user-role/entities/user-role.enti
 import { Role } from "src/usersCompanies/role/entities/role.entity";
 import { Boxsize } from "src/maintenence/boxes/entities/box.entity";
 import { comapny } from "src/auth/dto/create-auth.dto";
+import { randomUUID } from "crypto";
 
 @Injectable()
 export class NewCompanyService {
@@ -23,16 +24,28 @@ export class NewCompanyService {
     private userRepository: Repository<User>,
     @InjectRepository(UserCompany)
     private userCompanyRepository: Repository<UserCompany>,
-    @InjectRepository(UserCompany)
+    @InjectRepository(UsersRoles)
     private userRoleRepository: Repository<UsersRoles>,
     @InjectRepository(Boxsize)
     private boxSizeRepository: Repository<Boxsize>
   ) {}
   async create(createNewCompanyDto: CreateNewCompanyDto) {
-    if (createNewCompanyDto.AdminPassword !== "CmplAdmin")
+    if (createNewCompanyDto.AdminPassword !== "DannyCompulockyAdmin")
       throw UnauthorizedException;
-    ////////////////// CompanySetting /////////////////////////////
+    const comapnyExits = await this.companyRepository.findOne({
+      where: {companySetting:{priorityApiUrl : createNewCompanyDto.priorityApiUrl , 
+        priorityApiCompany: createNewCompanyDto.priorityApiCompany }}
+    })
+    if (comapnyExits)
+      throw new BadRequestException("Company already exists ", {
+          cause: new Error(),
+          description: "Company already exists",
+        });
+
+    ////////////////// CompanySetting /////////////////////////////  
+    console.log("start companySettingRepository");
     const cmpSetting = new CompanySetting();
+    
     cmpSetting.priorityApiCompany = createNewCompanyDto.priorityApiCompany;
     cmpSetting.priorityApiPassword = createNewCompanyDto.priorityApiPassword;
     cmpSetting.priorityApiUrl = createNewCompanyDto.priorityApiUrl;
@@ -50,6 +63,7 @@ export class NewCompanyService {
     const resCompanySetting = await this.companySettingRepository.save(
       cmpSetting
     );
+     console.log("end companySettingRepository");
     ////////////////// Company ////////////////////////////////////
     const cmp = new Company();
     cmp.name = createNewCompanyDto.priorityApiCompany;
@@ -58,39 +72,45 @@ export class NewCompanyService {
     cmp.companySetting = new CompanySetting();
     cmp.companySetting.id = resCompanySetting.id;
     const resCompany = await this.companyRepository.save(cmp);
+     console.log("end companyRepository");
     ////////////////// User ////////////////////////////////////
     const usr = new User();
     usr.userName = createNewCompanyDto.priorityApiCompany + "Admin";
     usr.userPasswordEnc = createNewCompanyDto.priorityApiCompany + "#Zbq";
-    usr.selectedCompany = cmp.id;
+    usr.userSurname = createNewCompanyDto.priorityApiCompany + "Admin";
+    usr.userUuid =  randomUUID();
+    usr.userMail = createNewCompanyDto.priorityApiCompany + 'Admin@mail.com';
+    usr.userMobile = '+001-';
+    usr.selectedCompany = resCompany.id;
     const resUser = await this.userRepository.save(usr);
+    console.log("end userRepository");
     ////////////////// UserCompany ////////////////////////////////////
     const usrCompany = new UserCompany();
-    usrCompany.users = new User();
-    usrCompany.users.id = usr.id;
-    usrCompany.company = new Company();
-    usrCompany.company.id = cmp.id;
-    const resUserCompany = await this.userCompanyRepository.save(usr);
+    usrCompany.users = { id: resUser.id } as User;
+    usrCompany.company = { id: resCompany.id } as Company;
+    const resUserCompany = await this.userCompanyRepository.save(usrCompany);
     ////////////////// UserRole ////////////////////////////////////
     const usrRole = new UsersRoles();
     usrRole.users = new User();
-    usrRole.users.id = usr.id;
+    usrRole.users.id = resUser.id;
     usrRole.role = new Role();
     usrRole.role.id = 7;
-    const resUserRole = await this.userRoleRepository.save(usr);
+    console.log(usrRole)
+    const resUserRole = await this.userRoleRepository.save(usrRole);
     ////////////////// BoxSize ////////////////////////////////////
     let boxSize = new Boxsize();
     boxSize.company = new Company();
-    boxSize.company.id = cmp.id;
+    boxSize.company.id = resCompany.id;
     boxSize.sizeDesc = "Pallet";
-    await this.boxSizeRepository.save(usr);
+    await this.boxSizeRepository.save(boxSize);
+    boxSize = new Boxsize();
     boxSize.company = new Company();
-    boxSize.company.id = cmp.id;
+    boxSize.company.id = resCompany.id;
     boxSize.sizeDesc = "Custom";
-    await this.boxSizeRepository.save(usr);
+    await this.boxSizeRepository.save(boxSize);
 
     return {
-      AdminUser: createNewCompanyDto.priorityApiCompany + "Admin",
+      AdminUser: createNewCompanyDto.priorityApiCompany + "Admin@mail.com",
       AdminPassword: createNewCompanyDto.priorityApiCompany + "#Zbq",
     };
 
@@ -109,27 +129,42 @@ export class NewCompanyService {
   //   return `This action updates a #${id} newCompany`;
   // }
 
-  async remove(id: string) {
+  async remove(id: string, AdminPassword: string) {
+     if (AdminPassword !== "DannyCompulockyAdmin")
+      throw UnauthorizedException;
     await this.boxSizeRepository.delete({
       company: { id: id },
     });
+    console.log("end boxSizeRepository");
     const users = await this.userCompanyRepository.find({
-      where: {company:{id:id}}
-    })
-    users.forEach(async (u)=>{
-      await this.userRoleRepository.delete({
-      users:{ id: u.id },
+      where: { company: { id: id } },
     });
-
-    })
-
+    users.forEach(async (u) => {
+      await this.userRoleRepository.delete({
+        users: { id: u.id },
+      });
+    });
+    console.log("end userRoleRepository");
     await this.userCompanyRepository.delete({
       company: { id: id },
     });
-    await this.companySettingRepository.delete({
-      company: { id: id },
-    });
+    console.log("end userCompanyRepository");
 
-    return await this.companyRepository.delete(id);
+    const resUser = await this.userRepository.delete({
+      selectedCompany : id
+    });
+  console.log("end userRepository");
+    const resComp = await this.companyRepository.findOne({
+      where: { id: id },
+    });
+    console.log("find userCompanyRepository", resComp);
+    const res = await this.companyRepository.delete(id);
+    console.log("end companyRepository");
+    if (resComp && resComp.companySetting ) {
+      await this.companySettingRepository.delete(resComp.companySetting.id );
+    }
+    console.log("end companySettingRepository");
+    
+    return res;
   }
 }
